templates,uiobjects,api,loader: move XML interpreter and CreateUIObject out of api.lua
Second phase of #838. api.lua was a dumping ground: it held CreateFrame
(genuine API surface), CreateUIObject (tightly coupled to the XML
interpreter, which lived in loader.lua), and object-graph primitives
(SetParent/ParentSub/DoSetParent/GetDebugName) used everywhere. Splitting
templates from closures to data needed api.lua to reach the interpreter,
but loader (which owned the interpreter) already depends on api -- so the
interpreter can't live in loader.lua if api needs to call it, without a
runtime-registered callback.

Resolution: recognize CreateUIObject and the XML interpreter are one
strongly-connected component (running a Kids phase can create child
uiobjects, which runs more phases) and put them in the same module, which
turns out to already be templates.lua (it owned the template registry).
uiobjects.lua -- previously just userdata/UserData -- picks up the object
graph primitives (SetParent/ParentSub/DoSetParent/GetDebugName), since
they're used by both templates.lua and points.lua and aren't part of the
cycle. api.lua shrinks to just CreateFrame/CreateChildUIObject/
CreateForbiddenFrame, delegating to templates.CreateUIObject. loader.lua
shrinks to addon discovery/load ordering/file reading, delegating to
templates.LoadFile; it no longer depends on api at all.

gencode.lua, cgencode.lua, mainloop.lua, runner.lua, points.lua, and the
handful of data/uiobjects and data/impl luafile snippets that referenced
api.CreateUIObject/api.SetParent/api.ParentSub/api.frames are repointed
to templates/uiobjects to match.

mkInitPhase still builds a bespoke closure per template element --
unchanged from before. Now that CreateUIObject and the interpreter are
co-located, the earlier "how does api.lua invoke a template's phases"
question is moot regardless of that choice, so it's left for a later,
lower-stakes pass.
