+++ data/impl/CreateFont.lua
@@ -1,9 +1,9 @@
-local api = ...
+local templates = ...
 local cache = {}
 return function(name)
   local font = cache[name]
   if not font then
-    font = api.CreateUIObject('font', name)
+    font = templates.CreateUIObject('font', name)
     cache[name] = font
   end
   return font
