+++ data/impl/EnumerateFrames.lua
@@ -1,5 +1,5 @@
-local api = ...
-local nextentry, arg = api.frames:entries()
+local templates = ...
+local nextentry, arg = templates.frames:entries()
 return function(frame)
   return nextentry(arg, frame)
 end
