+++ data/uiobjects/AnimationGroup/CreateAnimation.lua
@@ -1,8 +1,8 @@
-local api, uiobjecttypes = ...
+local templates, uiobjecttypes = ...
 return function(self, type)
   local ltype = (type or 'animation'):lower()
   if not (uiobjecttypes.Has(ltype) and uiobjecttypes.InheritsFrom(ltype, 'animation')) then
     ltype = 'animation'
   end
-  return api.CreateUIObject(ltype, nil, self)
+  return templates.CreateUIObject(ltype, nil, self)
 end
