+++ data/uiobjects/ModelScene/CreateActor.lua
@@ -1,5 +1,5 @@
-local api, templates = ...
+local templates = ...
 return function(self, name, template)
   local tmpls = template and { templates.GetTemplateOrThrow(template) }
-  return api.CreateUIObject('actor', name, self, nil, tmpls)
+  return templates.CreateUIObject('actor', name, self, nil, tmpls)
 end
