+++ data/uiobjects/ScrollFrame/SetScrollChild.lua
@@ -1,12 +1,12 @@
-local api, points = ...
+local points, uiobjects = ...
 return function(self, scrollChild)
   local old = self.scrollChild
   if old then
-    api.SetParent(old, nil)
+    uiobjects.SetParent(old, nil)
   end
   self.scrollChild = scrollChild
   if scrollChild then
-    api.SetParent(scrollChild, self)
+    uiobjects.SetParent(scrollChild, self)
     points.ClearAllPoints(scrollChild)
     points.SetPointInternal(scrollChild, 'TOPLEFT', self, 'TOPLEFT', 0, 0)
   end
